requirements.txt
pyspark>=3.5.0
torch>=2.0.0
transformers>=4.35.0
pandas>=2.0.0
pyarrow>=14.0.0
pytest>=7.0.0
kaggle>=1.6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
